Implement agent script tab changes check

+++ IqraCore/Entities/Business/App/Agent/Script/Node/BusinessAppAgentScriptNode.cs
@@ -11,7 +11,7 @@ namespace IqraCore.Entities.Business
 
     public class BusinessAppAgentScriptNodePosition
     {
-        public int X { get; set; } = 0;
-        public int Y { get; set; } = 0;
+        public double X { get; set; } = 0;
+        public double Y { get; set; } = 0;
     }
 }
